norminette ast
