+++ src/create_node.c
@@ -6,7 +6,7 @@
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/14 17:50:35 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/17 16:48:12 by jhor             ###   ########.fr       */
+/*   Updated: 2025/10/20 16:41:08 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,6 @@ t_token	*append_node(t_token *head, char *start, size_t end)
 {
 	t_token	*temp;
 	size_t	i;
-	t_token	*temp2;
 	t_token	*trave;
 	
 	i = 0;
@@ -53,9 +52,6 @@ t_token	*append_node(t_token *head, char *start, size_t end)
 	temp->lexeme = malloc(sizeof(char) * (end + 1));
 	if (!temp->lexeme)
 		return (NULL);
-	temp2 = NULL;
-	if (head->next != NULL)
-		temp2 = head->next;
 	while (i < end && start[i])
 	{
 		temp->lexeme[i] = start[i];
@@ -68,14 +64,6 @@ t_token	*append_node(t_token *head, char *start, size_t end)
 	while (trave->next != NULL)
 		trave = trave->next;
 	trave->next = temp;
-	// t_token *temp3 = NULL;
-	// temp3 = trave->next;
-	// while (temp3 != NULL)
-	// {
-	// 	printf("new node: %s\n", temp3->lexeme);
-	// 	temp3 = temp3->next;
-	// }
-	// printf("----------------------------------------------\n");
 	return (head);
 }
