+++ src/error_syntax.c
@@ -6,7 +6,7 @@
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/22 15:09:08 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/14 13:45:51 by jhor             ###   ########.fr       */
+/*   Updated: 2025/10/20 16:12:34 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,40 +36,40 @@ void	error_pipe(t_token *token)
 	return;
 }
 
-int	error_syntax(t_token *token)
-{
-	while (token != NULL)
-	{
-		if (*(token->lexeme) == '>' || *(token->lexeme) == '<'
-			|| ft_strncmp(token->lexeme, ">>", 2) == 0
-			|| ft_strncmp(token->lexeme, "<<", 2) == 0)
-		{
-			printf("current token lexeme is: %s\n", token->lexeme);
-			printf("current token enum is: %d\n", token->token);
-			if (token->next == NULL)
-			{
-				ft_putstr_fd("bash: syntax error near unexpected token `newline'\n", 2);
-				return (2);
-			}
-			else
-			{
-				token = token->next;
-				printf("current token lexeme is: %s\n", token->lexeme);
-				printf("current token enum is: %d\n", token->token);
-				printf("i am here in error handling redir\n");
-				error_redir(token);
-				return (2);
-			}
-		}
-		else if (*(token->lexeme) == '|')
-		{
-			if (token->next == NULL || token->next->token == PIPE)
-			{
-				error_pipe(token);
-				return (2);
-			}
-		}
-		token = token->next;
-	}
-	return (0);
-}
+// int	error_syntax(t_token *token)
+// {
+// 	while (token != NULL)
+// 	{
+// 		if (*(token->lexeme) == '>' || *(token->lexeme) == '<'
+// 			|| ft_strncmp(token->lexeme, ">>", 2) == 0
+// 			|| ft_strncmp(token->lexeme, "<<", 2) == 0)
+// 		{
+// 			printf("current token lexeme is: %s\n", token->lexeme);
+// 			printf("current token enum is: %d\n", token->token);
+// 			if (token->next == NULL)
+// 			{
+// 				ft_putstr_fd("bash: syntax error near unexpected token `newline'\n", 2);
+// 				return (2);
+// 			}
+// 			else
+// 			{
+// 				token = token->next;
+// 				printf("current token lexeme is: %s\n", token->lexeme);
+// 				printf("current token enum is: %d\n", token->token);
+// 				printf("i am here in error handling redir\n");
+// 				error_redir(token);
+// 				return (2);
+// 			}
+// 		}
+// 		else if (*(token->lexeme) == '|')
+// 		{
+// 			if (token->next == NULL || token->next->token == PIPE)
+// 			{
+// 				error_pipe(token);
+// 				return (2);
+// 			}
+// 		}
+// 		token = token->next;
+// 	}
+// 	return (0);
+// }
